Make naming conventions internals more consistent
This is stemmed from me noticing that most of these functions do not
mention Python or Ruby, as it is implied.  That may have to change
when RubyValues start creeping into Python and the whole thing may
just become really confusing without applying some method to the
madness.

For now, shorten things up.

Signed-off-by: Dan Farina <drfarina@acm.org>