Copy the passed text buffer instead of referencing it
It's not clear from the Python docs -- a rare miss -- but the
documentation for PyUnicode_FromString is as follows:

PyObject *PyUnicode_FromString(const char *u)
Return value: New reference.
Create a Unicode object from an UTF-8 encoded null-terminated char buffer u.

It doesn't mention if it's copied or not, but now I think it does not.
Or, I just hid a bug behind a change in access pattern, but given that
I am able to reproduce this right after printf-ing the bytes that
cause the unicode decode to blow up (and they look fine), I suspect
this is the real mechanism for the crash.

TODO: report that as a helpful clarification to Python.

Signed-off-by: Dan Farina <drfarina@acm.org>